Fold Booksleeve remaining test suite into main test project to de-dupe
