--- MigratedBookSleeveTestSuite/MigratedBookSleeveTestSuite.csproj
@@ -1,41 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <Description>MigratedBookSleeveTestSuite</Description>
-    <TargetFrameworks>net452;netcoreapp1.0</TargetFrameworks>
-    <AssemblyName>MigratedBookSleeveTestSuite</AssemblyName>
-    <PackageId>MigratedBookSleeveTestSuite</PackageId>
-    <GenerateDocumentationFile>false</GenerateDocumentationFile>
-    <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
-    <MaxParallelThreads>1</MaxParallelThreads>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\StackExchange.Redis\StackExchange.Redis.csproj" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.0.0" />
-
-    <!-- Temp workaround for analyzer issue -->
-    <PackageReference Include="xunit.analyzers" Version="0.6.1" />
-    <PackageReference Include="xunit" Version="$(xUnitVersion)" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="$(xUnitVersion)" />
-    <DotNetCliToolReference Include="dotnet-xunit" Version="$(xUnitVersion)" />
-  </ItemGroup>
-
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net452' ">
-    <Reference Include="System" />
-    <Reference Include="Microsoft.CSharp" />
-  </ItemGroup>
-
-  <PropertyGroup Condition=" '$(TargetFramework)' == 'netcoreapp1.0' ">
-    <DefineConstants>$(DefineConstants);PLAT_SAFE_CONTINUATIONS;CORE_CLR</DefineConstants>
-  </PropertyGroup>
-
-  <ItemGroup Condition=" '$(TargetFramework)' == 'netcoreapp1.0' ">
-    <PackageReference Include="System.Console" Version="$(CoreFxVersion)" />
-    <PackageReference Include="System.Diagnostics.Debug" Version="$(CoreFxVersion)" />
-    <PackageReference Include="System.Diagnostics.TraceSource" Version="$(CoreFxVersion)" />
-    <PackageReference Include="System.Linq.Expressions" Version="$(CoreFxVersion)" />
-    <PackageReference Include="System.Reflection.Extensions" Version="$(CoreFxVersion)" />
-    <PackageReference Include="System.Runtime.InteropServices" Version="$(CoreFxVersion)" />
-    <PackageReference Include="System.Threading.Tasks.Parallel" Version="$(CoreFxVersion)" />
-    <PackageReference Include="Microsoft.CSharp" Version="$(CoreFxVersion)" />
-  </ItemGroup>
-</Project>
