+++ StackExchange.Redis.Tests/Booksleeve/Config.cs
@@ -1,108 +1,17 @@
 ﻿using System;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Security.Authentication;
-using System.Threading.Tasks;
-using StackExchange.Redis;
 using Xunit;
 using Xunit.Abstractions;
-
-namespace Tests
-{
-    public class Config
-    {
-        public static string CreateUniqueName() => Guid.NewGuid().ToString("N");
-
-        internal static IServer GetServer(ConnectionMultiplexer conn) => conn.GetServer(conn.GetEndPoints()[0]);
-
-        static readonly SocketManager socketManager = new SocketManager();
-        static Config()
-        {
-            TaskScheduler.UnobservedTaskException += (sender, args) =>
-            {
-                Trace.WriteLine(args.Exception, "UnobservedTaskException");
-                args.SetObserved();
-            };
-        }
-
-        public const string LocalHost = "127.0.0.1"; //"192.168.0.10"; //"127.0.0.1";
-        public const string RemoteHost = "127.0.0.1"; // "ubuntu";
-
-        private const int unsecuredPort = 6379, securedPort = 6381,
-            clusterPort0 = 7000, clusterPort1 = 7001, clusterPort2 = 7002;
-
-        private ITestOutputHelper Output { get; }
-        public Config(ITestOutputHelper output)
-        {
-            Output = output;
-        }
-
-
-#if CLUSTER
-        internal static RedisCluster GetCluster(TextWriter log = null)
-        {
-            string clusterConfiguration =
-                RemoteHost + ":" + clusterPort0 + "," +
-                RemoteHost + ":" + clusterPort1 + "," +
-                RemoteHost + ":" + clusterPort2;
-            return RedisCluster.Connect(clusterConfiguration, log);
-        }
+#if !CORE_CLR
+using System.Security.Authentication;
 #endif
 
-        //const int unsecuredPort = 6380, securedPort = 6381;
-
-        internal static ConnectionMultiplexer GetRemoteConnection(bool open = true, bool allowAdmin = false, bool waitForOpen = false, int syncTimeout = 5000, int ioTimeout = 5000)
-        {
-            return GetConnection(RemoteHost, unsecuredPort, open, allowAdmin, waitForOpen, syncTimeout, ioTimeout);
-        }
-
-        private static ConnectionMultiplexer GetConnection(string host, int port, bool open = true, bool allowAdmin = false, bool waitForOpen = false, int syncTimeout = 5000, int ioTimeout = 5000)
-        {
-            var options = new ConfigurationOptions
-            {
-                EndPoints = { { host, port } },
-                AllowAdmin = allowAdmin,
-                SyncTimeout = syncTimeout,
-                SocketManager = socketManager
-            };
-            var conn = ConnectionMultiplexer.Connect(options);
-            conn.InternalError += (s, args) =>
-            {
-                Trace.WriteLine(args.Exception.Message, args.Origin);
-            };
-            if (open)
-            {
-                if (waitForOpen) conn.GetDatabase().Ping();
-            }
-            return conn;
-        }
-        internal static ConnectionMultiplexer GetUnsecuredConnection(bool open = true, bool allowAdmin = false, bool waitForOpen = false, int syncTimeout = 5000, int ioTimeout = 5000)
-        {
-            return GetConnection(LocalHost, unsecuredPort, open, allowAdmin, waitForOpen, syncTimeout, ioTimeout);
-        }
-
-        internal static ConnectionMultiplexer GetSecuredConnection(bool open = true)
-        {
-            var options = new ConfigurationOptions
-            {
-                EndPoints = { { LocalHost, securedPort } },
-                Password = "changeme",
-                SyncTimeout = 6000,
-                SocketManager = socketManager
-            };
-            var conn = ConnectionMultiplexer.Connect(options);
-            conn.InternalError += (s, args) =>
-            {
-                Trace.WriteLine(args.Exception.Message, args.Origin);
-            };
-            return conn;
-        }
-
-        internal static RedisFeatures GetFeatures(ConnectionMultiplexer muxer)
-        {
-            return GetServer(muxer).Features;
-        }
+namespace StackExchange.Redis.Tests.Booksleeve
+{
+    public class Config : BookSleeveTestBase
+    {
+        public Config(ITestOutputHelper output) : base(output) { }
 
         [Fact]
         public void CanOpenUnsecuredConnection()
@@ -117,7 +26,7 @@ namespace Tests
         [Fact]
         public void CanOpenSecuredConnection()
         {
-            using (var conn = GetSecuredConnection(false))
+            using (var conn = GetSecuredConnection())
             {
                 var server = GetServer(conn);
                 server.Ping();
@@ -130,10 +39,12 @@ namespace Tests
             Assert.Throws<RedisConnectionException>(() =>
             {
                 var log = new StringWriter();
-                try {
-                    using (var conn = ConnectionMultiplexer.Connect(Config.LocalHost + ":6500")) { }
+                try
+                {
+                    using (var conn = ConnectionMultiplexer.Connect(LocalHost + ":6500")) { }
                 }
-                finally {
+                finally
+                {
                     Output.WriteLine(log.ToString());
                 }
             });
@@ -162,7 +73,8 @@ namespace Tests
             var log = new StringWriter();
             try
             {
-                using (var conn = ConnectionMultiplexer.Connect(Config.LocalHost + ":6500,abortConnect=false")) {
+                using (var conn = ConnectionMultiplexer.Connect(LocalHost + ":6500,abortConnect=false"))
+                {
                     Assert.False(conn.GetServer(conn.GetEndPoints().Single()).IsConnected);
                     Assert.False(conn.GetDatabase().IsConnected(default(RedisKey)));
                 }
@@ -172,13 +84,15 @@ namespace Tests
                 Output.WriteLine(log.ToString());
             }
         }
+
         [Fact]
         public void CreateDisconnectedNonsenseConnection_DNS()
         {
             var log = new StringWriter();
             try
             {
-                using (var conn = ConnectionMultiplexer.Connect("doesnot.exist.ds.aasd981230d.com:6500,abortConnect=false", log)) {
+                using (var conn = ConnectionMultiplexer.Connect("doesnot.exist.ds.aasd981230d.com:6500,abortConnect=false", log))
+                {
                     Assert.False(conn.GetServer(conn.GetEndPoints().Single()).IsConnected);
                     Assert.False(conn.GetDatabase().IsConnected(default(RedisKey)));
                 }
@@ -188,6 +102,7 @@ namespace Tests
                 Output.WriteLine(log.ToString());
             }
         }
+
 #if !CORE_CLR
         [Fact]
         public void SslProtocols_SingleValue()
@@ -202,7 +117,7 @@ namespace Tests
         {
             var log = new StringWriter();
             var options = ConfigurationOptions.Parse("myhost,sslProtocols=Tls11|Tls12");
-            Assert.Equal(SslProtocols.Tls11|SslProtocols.Tls12, options.SslProtocols.Value);
+            Assert.Equal(SslProtocols.Tls11 | SslProtocols.Tls12, options.SslProtocols.Value);
         }
 
         [Fact]
@@ -213,7 +128,7 @@ namespace Tests
             // The below scenario is for cases where the *targeted*
             // .NET framework version (e.g. .NET 4.0) doesn't define an enum value (e.g. Tls11)
             // but the OS has been patched with support
-            int integerValue = (int)(SslProtocols.Tls11 | SslProtocols.Tls12);
+            const int integerValue = (int)(SslProtocols.Tls11 | SslProtocols.Tls12);
             var options = ConfigurationOptions.Parse("myhost,sslProtocols=" + integerValue);
             Assert.Equal(SslProtocols.Tls11 | SslProtocols.Tls12, options.SslProtocols.Value);
         }
@@ -222,7 +137,7 @@ namespace Tests
         public void SslProtocols_InvalidValue()
         {
             var log = new StringWriter();
-            Assert.Throws<ArgumentOutOfRangeException>(() => ConfigurationOptions.Parse("myhost,sslProtocols=InvalidSslProtocol"));            
+            Assert.Throws<ArgumentOutOfRangeException>(() => ConfigurationOptions.Parse("myhost,sslProtocols=InvalidSslProtocol"));
         }
 #endif
 
@@ -281,11 +196,6 @@ namespace Tests
             var options = new ConfigurationOptions();
             Assert.True(options.DefaultVersion.Equals(new Version(2, 0, 0)));
             Assert.True(options.AbortOnConnectFail);
-        }
-
-        internal static void AssertNearlyEqual(double x, double y)
-        {
-            if (Math.Abs(x - y) > 0.00001) Assert.Equal(x, y);
         }
     }
 }
