+++ StackExchange.Redis.Tests/Booksleeve/Issues/Issue10.cs
@@ -1,13 +1,16 @@
 ﻿using Xunit;
+using Xunit.Abstractions;
 
-namespace Tests.Issues
+namespace StackExchange.Redis.Tests.Booksleeve.Issues
 {
-    public class Issue10
+    public class Issue10 : BookSleeveTestBase
     {
+        public Issue10(ITestOutputHelper output) : base(output) { }
+
         [Fact]
         public void Execute()
         {
-            using (var muxer = Config.GetUnsecuredConnection())
+            using (var muxer = GetUnsecuredConnection())
             {
                 const int DB = 5;
                 const string Key = "issue-10-list";
@@ -20,9 +23,9 @@ namespace Tests.Issues
 
                 var contents = conn.Wait(conn.ListRangeAsync(Key, 0, -1));
                 Assert.Equal(3, contents.Length);
-                Assert.Equal("ghi", (string)contents[0]);
-                Assert.Equal("jkl", (string)contents[1]);
-                Assert.Equal("abc", (string)contents[2]);
+                Assert.Equal("ghi", contents[0]);
+                Assert.Equal("jkl", contents[1]);
+                Assert.Equal("abc", contents[2]);
             }
         }
     }
