+++ StackExchange.Redis.Tests/Booksleeve/Issues/SO11766033.cs
@@ -1,18 +1,21 @@
 ﻿using Xunit;
+using Xunit.Abstractions;
 
-namespace Tests.Issues
+namespace StackExchange.Redis.Tests.Booksleeve.Issues
 {
-    public class SO11766033
+    public class SO11766033 : BookSleeveTestBase
     {
+        public SO11766033(ITestOutputHelper output) : base(output) { }
+
         [Fact]
         public void TestNullString()
         {
             const int db = 3;
-            using (var muxer = Config.GetUnsecuredConnection(true))
+            using (var muxer = GetUnsecuredConnection(true))
             {
                 var redis = muxer.GetDatabase(db);
-                string expectedTestValue = null;
-                var uid = Config.CreateUniqueName();
+                const string expectedTestValue = null;
+                var uid = CreateUniqueName();
                 redis.StringSetAsync(uid, "abc");
                 redis.StringSetAsync(uid, expectedTestValue);
                 string testValue = redis.StringGet(uid);
@@ -24,11 +27,11 @@ namespace Tests.Issues
         public void TestEmptyString()
         {
             const int db = 3;
-            using (var muxer = Config.GetUnsecuredConnection(true))
+            using (var muxer = GetUnsecuredConnection(true))
             {
                 var redis = muxer.GetDatabase(db);
-                string expectedTestValue = "";
-                var uid = Config.CreateUniqueName();
+                const string expectedTestValue = "";
+                var uid = CreateUniqueName();
 
                 redis.StringSetAsync(uid, expectedTestValue);
                 string testValue = redis.StringGet(uid);
